+++ src/DotNetOutdated/PrereleaseReporting.cs
@@ -1,0 +1,9 @@
+﻿namespace DotNetOutdated
+{
+    internal enum PrereleaseReporting
+    {
+        Auto,
+        Always,
+        Never
+    }
+}