+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Display column headers ([#1](https://github.com/jerriep/dotnet-outdated/issues/1))
+- Allow specifying whether to include pre-release versions ([#4](https://github.com/jerriep/dotnet-outdated/issues/4))
 - Fix bug when latest version cannot be found ([#10](https://github.com/jerriep/dotnet-outdated/issues/10))
 
 ## [v0.1.0]